requirements.txt
django
powerline-status
pymysql
pytz
tzdata
sqlparse
cryptography
djangorestframework
djangorestframework-simplejwt
